Create setup script and make the extension installable
I have also renamed the lib directory to owncloudannex to make it
installable to as a python module.
