+++ owncloudannex/CommonFunctions.py
@@ -67,7 +67,7 @@ class BufferReader(io.BytesIO):
         if self._callback:
             try:
                 self._callback(*self._cb_args, **self._cb_kwargs)
-            except: # catches exception from the callback 
+            except: # catches exception from the callback
                 raise CancelledError('The upload was cancelled.')
         return chunk
 
@@ -504,7 +504,7 @@ def updateWanted(size, filetypes):
         expr += "("
         org_filetypes = re.compile("include=(.*?) ").findall(old_wanted)
         for t in filetypes:
-            expr += "include=*." + t + " or " 
+            expr += "include=*." + t + " or "
         expr = expr.strip()
         if expr.rfind(" ") > -1:
             expr = expr[:expr.rfind(" ")]
