+++ setup.py
@@ -1,0 +1,13 @@
+from distutils.core import setup
+
+setup(
+    name='owncloudannex',
+    version='',
+    packages=['owncloudannex'],
+    url='',
+    license='',
+    author='dopsi',
+    author_email='',
+    description='',
+    scripts=['git-annex-remote-owncloud']
+)
